+++ src/lib/hooks/useDebounceHook.ts
@@ -1,9 +1,11 @@
 ﻿import { useEffect, useRef, useState } from "react";
 import { TypeaheadOption } from "../types/Typeahead";
+import { getUniqueOptions } from "../helpers/typeahead";
 
 interface DebounceSearch {
   query: string;
   delay: number;
+  value: TypeaheadOption[];
 }
 
 const useDebounceHook = (
@@ -12,12 +14,12 @@ const useDebounceHook = (
   onQueryError?: ((error: unknown) => void) | undefined,
 ) => {
   const queryRef = useRef("");
-  const [loading, setLoading] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
   const [debounceSearch, setDebounceSearch] = useState<DebounceSearch | undefined>(undefined);
 
   useEffect(() => {
     if (debounceSearch) {
-      const { delay: counter, query } = debounceSearch;
+      const { delay: counter, query, value } = debounceSearch;
       queryRef.current = query;
 
       const timer =
@@ -37,18 +39,18 @@ const useDebounceHook = (
           try {
             const results = await queryFn(query);
             if (queryRef.current === query) {
-              setOptions(results);
+              setOptions(getUniqueOptions(results, value));
+              setIsLoading(false);
             }
           } catch (error) {
             onQueryError && onQueryError(error);
-          } finally {
-            setLoading(false);
+            setIsLoading(false);
           }
         })();
 
         setDebounceSearch(undefined);
       } else {
-        setLoading(true);
+        setIsLoading(true);
       }
 
       return () => {
@@ -58,9 +60,9 @@ const useDebounceHook = (
 
     // not all paths returns a value
     return undefined;
-  }, [queryFn, setOptions, debounceSearch, onQueryError]);
+  }, [queryFn, setOptions, onQueryError, debounceSearch]);
 
-  return { setDebounceSearch, loading };
+  return { setDebounceSearch, isLoading };
 };
 
 export { useDebounceHook };